feat: Initial commit for recipe website project
Created the main HTML file (index.html) for the recipe website project, defining the overall structure of the webpage. Started organizing the HTML structure to accommodate future content, such as recipe cards and additional pages. This commit lays the groundwork for building out the rest of the website's functionality and styling.